fix: separate fetch from field code from text control
